Create shim for tb.util.tensor_util (#1660)
tf.make_tensor_proto and tf.make_ndarray are going away. In order to
minimize the changes in one commit, we will first change all usage of
the tf tensor_util methods to point to tb.util.tensor_util before
we change the implementation.

Do note that when we implement the methods, we will have to be careful
with respect to proto definition we use. i.e., change most usage of
tf.Summary