Show scrollbar when content in the sidebar is tall
Previously, we had shrunken the content of the sidebar up to an extent
you cannot select any run. Instead, put a minimum height to the run
selector and now show a scrollbar to the sidebar so we can allow
taller content and shorter screen.

Manually inspected most plugins with shorter and taller screens.