Add template for PRs
* Motivation for features / changes

Reviewers of large PRs often do not have context to properly review or understand the motivation and supplied code changes. This template ensures that a bare minimum context is provided for the reviewer and commit reader.

* Technical description of changes

None, text only

* Screenshots of UI changes

None

* Detailed steps to verify changes work correctly (as executed by you)

None

* Alternate designs / implementations considered

This is the simplest process we could conceive of.