github: move issue and PR templates into subtree
Summary:
This commit simply reduces clutter at the top level.

Test Plan:
The [GitHub blog][1] indicates that this is a supported workflow, but
there’s no way to test it without merging this PR into the default
branch. After merging, draft (but don’t actually create) a new issue and
a new pull request, and ensure that the templates appear for each.

[1]: https://blog.github.com/2016-02-17-issue-and-pull-request-templates/

wchargin-branch: github-template-cleanup