test: add Travis test for “DO NOT SUBMIT” in files
Summary:
Google-internal tools perform this check, which is useful for preventing
premature merging of in-progress code. This commit adds equivalent
functionality to the Travis build only and not as a Bazel test,
analagous to our existing license test.

Test Plan:
Invoke the script and note that it passes with no output; revert #1716
and note that it fails with a helpful error message.

wchargin-branch: test-dns