test: add Travis test for “DO NOT SUBMIT” in files (#1717)
Summary:
Google-internal tools perform this check, which is useful for preventing
premature merging of in-progress code. This commit adds equivalent
functionality to the Travis build only and not as a Bazel test,
analagous to our existing license test.

Test Plan:
Invoke the script and note that it passes with no output. Run

    printf >hmmm 'DO NOT SUBMIT'

and note that the script fails with a helpful error message. Then run

    printf >hmmm 'DO NOT SUBMIT\0'

and note that the script passes because this is a binary file.

Then, run `(cd tensorboard && ./tools/do_not_submit_test.sh)`, and note
that the script fails because it is not being run from TensorBoard root.

wchargin-branch: test-dns