Use absl.flags instead of tf.flags (#1726)
TF is deprecating tf.flags. Migrate to absl.
Note that absl.flags by itself does not parse sys.argv like [1]
indicates. However, it is passed as part of `absl.app.run` [2] which we use,
so the change should have no behavioral difference.

Confirmed that the flags are parsed as expected by manually running the
demos and script/tools.

[1]: https://github.com/tensorflow/tensorflow/blob/master/tensorflow/python/platform/flags.py#L66
[2]: https://github.com/abseil/abseil-py/blob/master/absl/app.py#L292

Related to #1718.