Fix the relative path issue when Vulcanizing (#1738)
When specifying path to resources in, for instance, image tag, we use a
relative path. This path is supposed to be relative to the webpath (at
least when using WebfilesValidator) but vulcanized resources do not
convert this relative path into an appropriate path.

For instance, say we are trying to build index.html that refers to
another HTML file, foo/bar.html. In bar.html, we use a relative path
"./faz.png" to a resource in the same "foo" namespace. Previously, the
relative path was not expanded and was relative to the index.html. After
this change, we now expand faz.png based on foo/bar.html and make it
relative to output file index.html.