build: clean up, removing `--expunge`
Summary:
The previous `bazel clean --expunge` appears to have caused the build to
work. This commit removes the `--expunge` to check if running it _once_
was sufficient to unpoison some cache. If this build fails with a GCC
`-fcolor-diagnostics` error, then we’ll need to investigate more and/or
make a decision about whether we’re okay with keeping that in the build.
If this build passes, we’re probably okay _for now_, though this problem
may come back to bite us later.

Test Plan:
:shipit:

wchargin-branch: pin-tf-nightly