build: temporarily pin working tf-nightly version, and fix the build (#1758)
Summary:
This commit\* solves three problems.

First, the current-latest `tf-nightly` package is buggy (it does not
provide a `tf.compat.v1` module), so we pin an old version until a new
version with <https://github.com/tensorflow/tensorflow/pull/24867> is
released. That’s easy.

Second, the Travis build was hitting a numpy error that looks like
<https://github.com/numpy/numpy/issues/12736>. We started hitting this
now because numpy 1.16.0 was released, and something in the Travis
environment was causing a failure. The fix for this (included in this
commit) is to explicitly uninstall Travis’s preinstalled `numpy`.

Third, the Travis build was hitting a GCC error, wherein Bazel was
invoking `gcc` with the `-fcolor-diagnostics` argument, which Trusty GCC
does not support. This one is puzzling, as the functionality that Bazel
uses to test whether this argument is supported has not changed in a
while. It’s possible to reproduce the failure, but the circumstances
under which I could reproduce it are bizarre. Here is a full repro
script to be run on a fresh Ubuntu Trusty VM:

https://gist.github.com/wchargin/15ce5e31ab0976edca33eb4c0e6c36dc

Clearing Travis caches for `master` and this pull request through the
Travis UI does not fix the problem, but running `bazel clean --expunge`
does. I added that into the `.travis.yml` file, which caused builds to
succeed even after I removed the `--expunge` line (i.e., the cache only
needed to be expunged once). Suffice it to say that we have no idea
what’s happening here and are simply hoping that we’ve lucked out into
not having to deal with it again.

Test Plan:
Running bazel run `//tensorboard/pip_package:build_pip_package` includes
a smoke test that fails before this change and passes after it. This
addresses the first problem listed above. The other two problems can
only be tested by running on Travis.

wchargin-branch: pin-tf-nightly