ci: update Travis to Bazel 0.21.0
Summary:
Testing on fresh Trusty VMs on GCP suggests that this may suffice to
stop the strange errors wherein Bazel invokes GCC with invalid flags,
though it should be noted that we do not actually understand the root
cause or the mechanism of failure.

Test Plan:
Push to Travis; hope.

wchargin-branch: travis-bazel-0.21.0