manager: keep a `TensorboardInfo` database in /tmp
Summary:
As of this commit, TensorBoard instances will write “info files” to
`/tmp/.tensorboard-info/pid-${PID}.info` (where `/tmp` is a
platform-specific tempdir). The info files include the port number on
which the server is running and a description of the server’s
configuration, so that clients can check whether a suitable TensorBoard
instance exists (and, if so, on which port) before starting a new one.

A typical info file looks like this:

    $ cat /tmp/.tensorboard-info/pid-92239.info
    {
        "cache_key": "<long base64 string; omitted>",
        "db": "",
        "logdir": "/usr/local/google/home/wchargin/tensorboard_data/",
        "path_prefix": "",
        "pid": 92239,
        "port": 7012,
        "start_time": 1549064188,
        "version": "1.13.0a0"
    }

Test Plan:
Launch two TensorBoard instances. Then, comment out the `tb_logging`
import in `tensorboard/manager.py`, and open a Python shell and run

    from tensorboard import manager
    infos = manager.get_all()

Note that `len(infos) == 2` and the info objects are correct. Then, send
SIGINT to one of the TensorBoards and SIGTERM to the other, and note
that each of them has its info file removed and that `get_all()` returns
an empty list.

wchargin-branch: write-tensorboardinfo