main: handle SIGTERM, running `atexit` handlers (#1802)
Summary:
Currently, TensorBoard does not change the default signal dispositions
for any signal handlers, so if killed by SIGTERM or SIGQUIT it will exit
without running its `atexit` handlers, such as the one registered by the
DB import mode. As of this commit, we handle SIGTERM by exiting
gracefully. We leave SIGQUIT at the default disposition (temporary files
will not be cleaned up), in accordance with the GNU libc guidelines:
<https://www.gnu.org/software/libc/manual/html_node/Termination-Signals.html#index-SIGQUIT>

The implementation is not perfect. Ideally, we would perform our
graceful cleanup and then kill ourselves with the same signal to
properly inform our parent of the source of the exit: for details, see
<https://www.cons.org/cracauer/sigint.html>. But `atexit` doesn’t
provide a function like “please run all registered handlers now but
don’t actually quit”; we might be able to implement this in Python 2.7
using `sys.exitfunc`, but that’s deprecated in Python 2.7 and removed in
Python 3. If we want to do this right, we could implement our own
version of `atexit` (which would not be hard: the module is tiny). For
now, I’m comfortable with submitting this mostly-correct patch.

Supersedes part of #1795.

Test Plan:
Run `bazel build //tensorboard` and add the built binary to your PATH.
Then, run `tensorboard --logdir whatever && echo after`, wait for it to
print the “serving” message, and send SIGTERM via `kill(1)`. Note that
TensorBoard prints a message and exits cleanly, and that “`after`”
is printed to the console.

Patch `program.py` to add a pre-existing signal handler and `atexit`
cleanup from the top of `main`:

```diff
diff --git a/tensorboard/program.py b/tensorboard/program.py
index da59b4d1..c07cb855 100644
--- a/tensorboard/program.py
+++ b/tensorboard/program.py
@@ -201,6 +201,14 @@ class TensorBoard(object):

     :rtype: int
     """
+    def fake_handler(signum, frame):
+      print("Handling some signals...")
+    assert signal.signal(signal.SIGTERM, fake_handler) == signal.SIG_DFL
+    def fake_cleanup():
+      print("Cleaning everything up...")
+    import atexit
+    atexit.register(fake_cleanup)
+
     self._install_signal_handler(signal.SIGTERM, "SIGTERM")
     if self.flags.inspect:
       logger.info('Not bringing up TensorBoard, but inspecting event files.')
```

Then, re-run the above steps, and note that the signal handler and
`atexit` handler are both executed prior to cleanup:

```
$ tensorboard --logdir whatever && echo after
TensorBoard 1.13.0a0 at <hostname> (Press CTRL+C to quit)
TensorBoard caught SIGTERM; exiting...
Handling some signals...
Cleaning everything up...
after
```

Ideally, `after` should _not_ be printed; that it is is a consequence of
the fact that we don’t properly propagate the WIFSIGNALED flag as
described above.

wchargin-branch: handle-signals