manager: implement `start` (#1809)
Summary:
This function starts a new TensorBoard process with the given arguments,
or reuses an existing compatible process. It returns a `TensorboardInfo`
object describing how to reach the resulting TensorBoard process
(whether new or reused). See docs for more details.

Test Plan:
End-to-end tests included. These appear to be lightly flaky: I ran

    bazel test //tensorboard:manager_e2e_test --runs_per_test=100

six times on each of Python 2 and 3, and experienced three total
failures on Python 2 and zero on Python 3. On my machine, the test takes
14.7±0.9s to run on Python 2, and 17.9±1.0s to run on Python 3.

To test manually, run `bazel build //tensorboard`, then add that binary
to your path and head over to a Python REPL:

    $ export PATH="$(readlink -e ./bazel-bin/tensorboard):$PATH"
    $ python
    >>> from tensorboard import manager
    >>> r1 = manager.start(["--logdir", "~/tensorboard_data", "--port", "0"])
    >>> type(r1)
    <class 'tensorboard.manager.StartLaunched'>
    >>> r2 = manager.start(["--logdir", "~/tensorboard_data", "--port", "0"])
    >>> type(r2)
    <class 'tensorboard.manager.StartReused'>
    >>> r1.info == r2.info
    True
    >>> r1.info.port
    39081
    >>> import os
    >>> os.system("curl --silent localhost:39081 | tail -c 64")

    <tf-tensorboard use-hash brand="TensorBoard"></tf-tensorboard>
    0
    >>> manager.get_all() == [r1.info]
    True
    >>> os.kill(r1.info.pid, 15)
    >>> manager.get_all() == []
    True

wchargin-branch: manager-start