test: resolve TensorBoard binary relative to test
Summary:
Within Google, the `//tensorboard:tensorboard` target is at a different
path in the build workspace, so tests with a data dependency on it
cannot find it at `./tensorboard/tensorboard`. However, the binary
should always be in the same place relative to the test binary, so we
now use that instead.

Test Plan:
Running `bazel test //tensorboard:manager_e2e_test` still works in
Python 2 and 3.

wchargin-branch: relative-data-path