notebook: add Colab integration shell
Summary:
This hooks up the `notebook` module to Colab. The Colab runtime
sandboxes the contents of the output frame, but provides a service
worker tunnel so that the frame may communicate with underlying VM. We
take advantage of this by loading TensorBoard through a JavaScript shell
that changes the document `baseURI` to proxy requests through this
worker.

Test Plan:
Add `//tensorboard:notebook` to the deps of `build_pip_package`, then
build the Pip package and manually upload it onto a public Colab
instance. Then run:

    !pip uninstall -q -y tensorflow tf-nightly-2.0-preview
    !pip install -q tf-nightly-2.0-preview
    !pip uninstall -q -y tensorboard
    !pip install -q ./tensorboard-1.13.0a0-py3-none-any.whl

    %load_ext tensorboard.notebook
    %tensorboard --logdir ./tensorboard_data/mnist

and watch TensorBoard come to life as desired.

wchargin-branch: notebook-colab-integration