pip: add `tensorboard.notebook` to Pip package
Summary:
Users of the Pip package now have the files necessary to use Jupyter
extensions, by incanting `%load_ext tensorboard.notebook` at the top of
their notebook.

This does not add a `notebook` attribute to the root `tensorboard`
module; in this respect, it is like `backend`, not `program`.

Test Plan:
Run `bazel run //tensorboard/pip_package:build_pip_package`, and suppose
that `${WHEEL}` is the path to the emitted Python 3 wheel. Then run

    $ cd "$(mktemp -d)"
    $ . ./ve/bin/activate
    (ve) $ pip install -q tf-nightly-2.0-preview
    (ve) $ pip uninstall -q -y tb-nightly
    (ve) $ pip install -q "${WHEEL}"
    (ve) $ pip install -q jupyter
    (ve) $ jupyter notebook

Then, create a notebook and execute

    %load_ext tensorboard.notebook
    %tensorboard --logdir ~/tensorboard_data/mnist --port 0

to see a new TensorBoard instance.

wchargin-branch: tensorboard-notebook-in-pip