lazy: properly handle errors during load
Summary:
If a module fails to load once, we propagate its raised error. If the
module fails to load again, we mistakenly think that we’re still
executing the first load, so we raise the _wrong_ error the second time.
This patch fixes that.

Test Plan:
Unit test added. Running `bazel test //tensorboard:lazy_test` passes,
but fails if you revert the change to `lazy.py`.

wchargin-branch: lazy-errors