ui: open help in new window/tab (#1836)
Summary:
In a normal standalone TensorBoard, users probably want the help button
to open in a new tab rather than replacing their current state.

In a Jupyter or Colab notebook context, this is more important, because
the GitHub response’s CSP includes `frameancestors 'none'`.
Consequently, clicking the “help” button causes the frame contents to be
erased entirely, replaced with a white page. The user can right-click
and select “Back” to go back to TensorBoard, but this is far from
obvious, and also bad.

Test Plan:
Build TensorBoard, put it on your path, and run `jupyter notebook`. Then
use a `%tensorboard` command to launch TensorBoard, and verify that the
help link properly opens the documentation in a new tab.

wchargin-branch: ui-help-blank