scalars: use `tf-downloader` for download links
Summary:
We have a `<tf-downloader>` element that provides CSV and JSON download
links alongside a run selector, but this element is actually unused.
This commit replaces the scalar plugin’s reimplementation, updating the
styling of `<tf-downloader>` to stay mostly consistent with the existing
UI. This will enable us to fix #1845 in a way that can work for other
download links, too.

Test Plan:
Observe that the UI is virtually identical (the only visible change is
that the vertical position of the dropdown menu has shifted slightly):

![“Before” and “after” screenshot.](https://user-images.githubusercontent.com/4317806/52752587-38795e00-2fa8-11e9-8b7c-1d8b9a540e49.png)

Check that the download links download identical files before and after
the change, with the same filenames.

wchargin-branch: scalars-tf-downloader