Graph plugin populates graph and RunMeta v2.
This change pipes more data into the graph plugin that is persisted
on the TensorFlow's side with the new summary ops.

Also refactored the test a bit but it is hard to test anything
because the nightlies do not contain the required changes yet.
In a follow up changes, we will add test using more public APIs.