Scalars tutorial for TensorBoard and Keras  (#1876)
This is a very quick introduction to viewing default scalar summaries when developing a Keras model in TensorFlow. It's intended to be very simple to show that virtually no extra code is required to get started using TensorBoard to see how loss is changing during training.

To some extent this repeats some of the material in the "Quick Start", but since it's very focussed I don't think the repetition is a bad thing.