notebook: change default height to 800px
Summary:
The previous default height of 600px is okay for some dashboards but
particularly cramped for hparams. Through rigorous testing, we have
determined that 800px is a larger number.

Before and after screenshots of the hparams dashboard:
![Screenshots](https://user-images.githubusercontent.com/4317806/53537097-aac86300-3abd-11e9-8f98-f202a6191ce3.png)

Test Plan:
Build a Pip package, upload it to Colab, run the hparams demo, take
screenshots above.

wchargin-branch: notebook-enlarge-default-height