Make tests runnable with TF v2.
This change does not fix all the tests to be meaningfully
run with TensorFlow v2. Instead, it adds a decorator that let
one decorate a class or a func that are not yet v2 compliant.
The ultimate goal is to write a v2 variant that tests the v2
in a meaningful way.