profile: update demo data to include trace-only run (#1942)
Summary:
We make this change because we have a bug in profile plugin that will be
triggered if logdir only contains trace data.

Test Plan:
Remove any old profile demo data, then regenerate the data and launch
TensorBoard:

```
$ rm -rf /tmp/profile_demo/
$ bazel run //tensorboard/plugins/profile:profile_demo
$ bazel run //tensorboard -- --logdir /tmp/profile_demo
```

Prior to this commit, this would have _appeared_ to work; now, it hangs
on the “Processing datasets” screen. Fix to come shortly.