Make graph plugin usable without Graph data
Graph plugin is possible to use without any graph data on event file by
using its file selector. However, the left panel was hidden when the
event file is missing and thus not possible.

While making the change, I decided to update some pointer to v1 docs for
the v2 preparedness. If TensorBoard team decides to maintain v1.x branch
for 1.14+, we may want a different copy.

This change addresses #1613.