Test against TF2 preview in CI (#1911)
This change does not fix all the tests to run with TensorFlow v2.
Instead, it adds a decorator that let one decorate a class or a
func that are not yet v2 compliant. The ultimate goal is to write a
v2 variant that tests the v2 where necessary over time.