Add UMAP to embedding projector (#1901)
* Motivation for features / changes

Uniform Manifold Approximation and Projection ([UMAP]) is a dimension
reduction technique that can be used for visualisation similarly to
t-SNE. I've ported the python implementation to JavaScript to add an
additional projection method to the embedding projector.

The UMAP projection technique pretty much follows the approach for the
tSNE projection. The UMAP code is grouped in its own directory for ease
of maintenance. There are two things to note about this current
implementation of UMAP in the embedding projector:

1) UMAP's SGD optimization doesn't really animate as nicely as t-SNE.
I've addresses this by choosing not to animate the steps and just track
progress through a dialog.

2) I'm currently reusing the GPU computed k-nearest neighbors - there
didn't seem to be too big of a difference between using the UMAP CPU knn
approach vs the existing GPU approach in the projector. To note, the GPU
knn approach uses cosine distance, while the rest of UMAP's
implementation uses euclidean distance. Again, the subjective results
seem fine with this choice.

[UMAP]: https://github.com/lmcinnes/umap

* Screenshots of UI changes

![image](https://user-images.githubusercontent.com/7818584/52659999-88c2c400-2eb3-11e9-9b68-eab380cbf11d.png)

* Detailed steps to verify changes work correctly (as executed by you)

Run the [tf-dev-summit-tensorboard-tutorial] demo to generate test data
with embeddings, then point TensorBoard at that log directory and
exercise the new UI functionality in the UMAP section of the projector
dashboard.

[tf-dev-summit-tensorboard-tutorial]: https://github.com/martinwicke/tf-dev-summit-tensorboard-tutorial