graph: fixed most visual regressions of tf-graph
Nature of the changes: with Polymer upgrade, we now use the real
shadowDOM and our app broke because of that. Some of the changes are:
1. Incorrect selector: instead of global/element.querySelector, we now
use the proper Polymer API or use querySelector on a shadowRoot.
2. CSS changes: "::content" is deprecated.
3. CSS stylesheet is mostly sandboxed and changed document.styleSheets
to copy only the necessary portions.

Functional break that was not addressed:
1. node toggle is broken
2. fit to screen is broken
3. tf-graph-controls' legend is not rendering correctly.
4. icons (for the minimap, legends, info panel) are broken

In subsequent changes, we will address them.