graph: fixed tf-graph-icons and moved it for reusability
In the polyfilled WebComponent world where no DOM is properly scoped
allowed SVG definitions to be declared in one part of the DOM and reused
anywhere. Using this property, tf-graph-scene declared the SVG
definitions for nodes and edges while other components like legend,
info-panel, and the sidebar were using them.

Now that we can no longer do that, we decided to create a component used
by most of above components that define the gradient/node/edges where it
is used. This does create many duplicate of the definitions but does not
seem to have negative impact on runtime performance.