beholder: fix image resizing
Summary:
This reverts part of #1734 to fix #2073. A change to the TensorFlow APIs
added a static shape requirement, but Beholder’s image resizer operates
on images whose size is not statically known.

Test code adapted from @comzyh’s MWE in #2073—thanks!

Test Plan:
Added a unit test, which has the correct pass/fail behavior on current
`tf-nightly` and TensorFlow 1.13.1. The test does not run in TF 2.0,
because Beholder depends on sessions.

wchargin-branch: fix-2073