Revert "Add UMAP to embedding projector (#1901)"
Summary:
This reverts commit 37c5b50108be97e7bc48ef9d646cb8b52302d472. Manual
merge of `third_party/js.bzl` was required due to #2079; the file is now
identical to its state prior to #1901.

This is a temporary measure to keep our GitHub-to-Google sync process
green while we work on integrating the `umap-js` library into google3.

Test Plan:
The embeddings projector now lacks a UMAP option, but still works.

wchargin-branch: revert-1901