graph: sync parser config to proto counterpart (#2068)
A repeated field in Graph proto definition requires an entry in the
parser.ts to parse the text proto correctly. It was not only missing
entries for FunctionDefLibrary but also missed ones for few properties.

Do note that we did not encounter issuses because we do not make use of
most of the properties.