Support browser fetch with retry/queue
RequestManager is a thin wrapper around XMLHttpRequest and fetch.
Unfortunately, for some convenience of plugin authors, XMLHttpRequest
version hide away certain control over the request or response. For
fetch API, we would like to keep it as true to the real browser API
while giving it feature of automatic retries and request queueing.