Fixed histogram with the new category-paginated-view
This change improves the tf-paginated-view to work around the issue.
Currently tf-category-pane and tf-paginated-view together determine
whether a child is visible or "active". The problem is that in canonical
polymer component structure, it is difficult to pass information
"downwards" when the common ancestor is a host component. Without an
easy way, the proper way is for the host component,
tf-histogram-dashboard to collect the information and pass it down to
the respective tf-histogram-loader. This pattern was implemented but
caused nontrivial bloat that had to be repeated everywhere that used
tf-category-pane and tf-paginated-view. There are rough two ways to
solve this issue: (1) mixin and (2) dom-repeat like component.

We decided to use (2) by improving our tf-paginated-view2 experiment.