Fixed Plottable CSS issue
In order for Plottable based charts to properly work, we need to load
the plottable.css. Before, when using shadyCSS, including plottable.css
in the plottable.html was sufficient but now that we have the real
shadow CSS, we need to embed the CSS (styles) in respective components
using `<style include="componentName"></style>`.