profile: fixed hosts not updating issue.
When active hosts update, due to the discrepency with v1 version of
paper-dropdown-menu and paper-listbox, the dropdown does not react to
any changes in the listbox DOM/values changes.

To be honest here, I expect more part of the apps to be broke with this
behavioral change. We will fix them one by one.