Check if tensor value only has 1 shape
Previously, if a tensor value only has 1 shape, graph plugin logic would
fail to parse the list of shapes because pbtxt-parsing logic does not
differentiate between a field with a single message or a repeated field
containing a single message. This prevented an internal graph from being
parsed (with a console error).

Now, we can verify that the graph is rendered.