fixup: document `Discrete` init; use `TypeError`
Summary:
The `Domain` initializer has some slightly non-obvious semantics that
should really be documented.

The appropriate error class is not entirely clear-cut. I had originally
used `ValueError`, because this is what the docstrings unambiguously
recommend:

    >>> print(ValueError.__doc__)
    Inappropriate argument value (of correct type).
    >>> print(TypeError.__doc__)
    Inappropriate argument type.

If the argument is `[1, "two"]`, then the argument type is `list`, which
is appropriate; it is the argument’s _value_ (specifically, the type of
its second element) that is inappropriate. But I’m inclined to interpret
these more liberally, as if `values[i]` were a bone fide argument.

For instance, how would these rules address splat parameters? Given a
function definition `def foo(*x)` and a call site `foo(7)`, at the call
site the argument is clearly `7`, but in the function body the argument
is clearly `[7]`, and so the literal interpretation of these error types
is hopelessly inconsistent.

Test Plan:
Tests amended, and still pass on both Python 2 and Python 3.

wchargin-branch: hparams-structured-api