Refactor graph-loader to factor out graph loading
tf-graph-control, being inside graph plugin, changed recently to work
better with other graph components like tf-graph-controls. More
specifically, the problem happen because the tf-graph-loader was
refactored to take `selection` as the primary input to the loader and
it used run & type of graph selected by user to fetch appropriate data.
This contrasts with previous input `datasets` which took a `path` as an
input which loader used to fetch the data from. tf-graph-app and
tf-debugger-dashboard plugin were using the datasets prop to pass in
the special endpoint where the data lives.

Before describing refactor, we should define what the tf-graph-loader
should be. The loader is a UI-less functional component (renders no DOM;
uses lifecycle method and props to provide some functionality) that is
capable of fetching various informations like graph, conceptual graph,
and profile info (RunMetadata). We have recently introduce more
complexity to it.

Interestingly, all tf-graph variants (internal to Google too) does not
make use of any other data type other than the graph and there is almost
no reason to use the tf-graph-loader. The graph utilities provide an
ability to fetch and parse GraphDef, create SlimGraph, and create a
hierarchy with ease. The core values the tf-graph-loader provides are
1. error handling and 2. progress tracker (uses Polymer API). By
reusing the component, one can provide some consistency there.

Here, in this case, we decided to put as much load+parsing into utility
method and have loader specific to the dashboard while having another
loader that is API compatible with old tf-graph-loader.