Compile TypeScript to ES6
Background: Projector plugin makes heavy use of ES6 class but due to the
implementation changes in Polymer to use ES6 class, transpilation became
a bit more problematic. More specifically, because class in TypeScript
is transpiled into ES5 class syntax, Polymer class which is still in ES6
class had to interop with the ES5 way of inheritance. Empirically, it
seems to badly break.

Because we use Closure compiler on top of TSC, the ES6 class will get
transpiled by the Closure compiler.