custom scalars: fixed chart loading and rendering
Custom scalars does not use the paginated view and thus
we cannot apply the same tf-category-paginated-view. Using only
category-pane, the custom scalars has an issue of mounting
component/fetching data even when a component is not visible. Though
often innocuous, this poses issues for rendering. Plottable, basing
layout on the container size, cannot properly discern the size of the
figure and render squished chart. This change basically instructs the
chart to re-render when its parent's open state changes.