hparams: add (mostly-)notf `hparams_config_pb` (#2174)
Summary:
We reimplement `hparams_config` in terms of a new `hparams_config_pb`,
which has no direct dependencies on TensorFlow. The `metadata` module
unfortunately still emits TensorFlow protos, so there is still a
transitive dependency.

Note that this duplicates a small bit of logic from `summary`. We don’t
reimplement `summary` in terms of `summary_v2`, because we intentionally
provide a smaller API in `summary_v2` (e.g., we omit experiment name and
owner, which are unused). It is thus not possible to implement the full
functionality of `summary` in terms of `summary_v2`.

Test Plan:
Existing unit tests suffice for coverage, and also updated to test the
new endpoint explicitly.

wchargin-branch: hparams-config-pb