test: skip V1-only tests under `TF2_BEHAVIOR`
Summary:
This fixes some Google-internal issues (see <http://cl/247247363>). It
should be a no-op externally; we do not set this environment variable.

I believe that this resolves #1996 to the best of our ability without
relying on private TensorFlow symbols. The exact answer is given by

    importlib.import_module("tensorflow.python.tf2").enabled()

but we do not want to depend directly on the `tensorflow.python.*`
modules.

Test Plan:
Count the number of skipped tests before and after this change:

```
tests="$(bazel query "
    tests(rdeps(
        //tensorboard/...,
        set($(git grep -l run_v1_only)) - //tensorboard/util:test_util.py
    ))
")" &&
bazel test ${tests} --test_output=all 2>&1 |
    sed -n -e 's/.*(skipped=\([0-9]\+\)).*/\1/p' |
    awk '{ c += $1 } END { print c }'
```

Note that it prints `166` both times.

wchargin-branch: respect-tf2-behavior