profile: standardize tutorial style
Summary:
Per @lamberta via [the Google dev docs style guide][guide], we should
use sentence case for titles and headers, and “prefer italics instead
bold for emphasis” because “bold style tends to conflict with the
h3/h4/h5 styles that devsite uses when skimming the page”.

This commit replaces strong emphasis with emphasis, quotations, or code
spans, as appropriate. Strong emphasis is retained for description list
items and a `**NOTE**:` specifier (changed to `**Note:**` for
consistency with [the same guide][guide]).

This commit is easier to review with `--word-diff`.

[guide]: https://developers.google.com/style/capitalization

Test Plan:
Check that all remaining `**`-sequences correspond to description lists:

```
$ grep -F '**' docs/r2/tensorboard_profiling_keras.ipynb | cut -c -40
        "*   **CPU**: CPU events are und
        "*   **GPU**: GPU events are und
        "*   **TensorFlow Runtime**: Run
        "**NOTE**: Please don't run mult
```

wchargin-branch: profile-unembolden